convert to jsbundling
