--- app/assets/javascripts/application.js
@@ -1,41 +1,0 @@
-// This is a manifest file that'll be compiled into application.js, which will include all the files
-// listed below.
-//
-// Any JavaScript/Coffee file within this directory, lib/assets/javascripts, vendor/assets/javascripts,
-// or vendor/assets/javascripts of plugins, if any, can be referenced here using a relative path.
-//
-// It's not advisable to add code directly here, but if you do, it'll appear at the bottom of the
-// compiled file.
-//
-// Read Sprockets README (https://github.com/sstephenson/sprockets#sprockets-directives) for details
-// about supported directives.
-//
-//= require jquery3
-//= require 'honeybadger'
-
-//= require rails-ujs
-//= require turbolinks
-//
-// Required by Blacklight
-//= require blacklight/blacklight
-
-//= require popper
-//= require twitter/typeahead
-//= require bootstrap
-
-//= require bibliography
-//= require cited_documents
-//= require blacklight_gallery/blacklight-gallery
-//= require blacklight_oembed
-//= require full_text_collapse
-//= require index_status_typeahead
-//= require exhibit_search_typeahead
-//= require nested_related_items
-//= require openseadragon
-//= require openseadragon-rails/openseadragon-rails
-//= require spotlight
-//= require L.Control.Sidebar
-//= require blacklight_heatmaps
-//= require table_of_contents
-//= require site_search_type_toggle
-//= require blacklight_heatmaps_overrides
