--- app/assets/javascripts/blacklight_heatmaps.js
@@ -1,1 +1,0 @@
-//= require blacklight_heatmaps/default
