+++ app/javascript/blacklight.js
@@ -1,0 +1,2 @@
+import Blacklight from "blacklight-frontend";
+window.Blacklight = Blacklight
