+++ app/javascript/blacklight_heatmaps.js
@@ -1,0 +1,6 @@
+import 'blacklight-heatmaps/vendor/assets/javascripts/leaflet_solr_heatmap'
+import BlacklightHeatmaps from 'blacklight-heatmaps/app/assets/javascripts/blacklight_heatmaps/blacklight_heatmaps'
+import 'blacklight-heatmaps/app/assets/javascripts/blacklight_heatmaps/basemaps'
+import 'blacklight-heatmaps/app/assets/javascripts/blacklight_heatmaps/icons'
+import 'blacklight-heatmaps/app/assets/javascripts/blacklight_heatmaps/viewers/index'
+import 'blacklight-heatmaps/app/assets/javascripts/blacklight_heatmaps/viewers/show'
