+++ app/javascript/blacklight_heatmaps_overrides.js
