+++ app/javascript/blacklight_oembed.js
