+++ app/javascript/full_text_collapse.js
