+++ app/javascript/image_viewer_state_modal.js
