+++ app/javascript/select_image_area.js
