+++ app/javascript/sir_trevor.js
@@ -1,0 +1,2 @@
+import SirTrevor from 'sir-trevor'
+window.SirTrevor = SirTrevor
