+++ app/javascript/sir_trevor_block_overrides.js
