+++ app/javascript/spotlight.js
