+++ app/javascript/table_of_contents.js
